requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
pyinaturalist>=0.19.0
requests>=2.28.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
PyYAML>=6.0
Flask>=3.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.3.0
